requirements.txt
Flask==3.1.2
gunicorn==20.1.0
openai==1.100.2
faiss-cpu==1.12.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.9
huggingface-hub==0.34.4
sentence-transformers==5.1.0
scikit-learn==1.7.1
numpy==2.3.2
python-dotenv==1.1.1
